SUBJ: Recording scores MSG: 1) Finally add the game timer to the game flow, start the timer upon game start, end the timer upon game end 2) Check for win every game tick. If there is a winner, stop the match, record the score. 3) CSV R/W shit WIP: 1) My paths for everything are total ass. 2) High scores page (working on next)

+++ HighScoreManager.java
@@ -24,7 +24,7 @@ public class HighScoreManager {
     }
 
     public static boolean
-    addHighScore(CSVWriter writer, String name, int score) {
+    addHighScore(CSVWriter writer, String name, double score) {
         if (writer == null || name == null || score < 0)
             throw new IllegalArgumentException();
         String[] colsFinal = {name, String.valueOf(score),
