+++ PongTimer.java
@@ -2,18 +2,18 @@
  * Created by coxjc on 12/4/16.
  */
 public class PongTimer {
-    private long startMS = 0L; //Current timestamp in MS
-    private long finalMS = 0L; //Final timestamp in MS
-    private long elapsedMS = 0L; //MILLISECONDS elapsed since start of timer
+    private int startMS = 0; //Current timestamp in MS
+    private int finalMS = 0; //Final timestamp in MS
+    private int elapsedMS = 0; //MILLISECONDS elapsed since start of timer
 
     //Instantiates a timer, sets initial MS
     public void startTimer() {
-        this.startMS = System.currentTimeMillis();
+        this.startMS = (int) System.currentTimeMillis();
     }
 
     //Ends a timer, sets final MS, makes call to set elapsed seconds
     public void endTimer() {
-        this.finalMS = System.currentTimeMillis();
+        this.finalMS = (int) System.currentTimeMillis();
         this.setElapsedMS();
     }
 
@@ -23,7 +23,7 @@ public class PongTimer {
     }
 
     //Returns elapsed milliseconds
-    public long getElapsedMS() {
+    public int getElapsedMS() {
         return this.elapsedMS;
     }
 }
